TaskService (initial, UI, persistence, linter)

+++ arwo-core/src/main/java/io/github/greyp9/arwo/core/lang/MathU.java
@@ -41,7 +41,7 @@ public final class MathU {
         return l;
     }
 
-    public static int log(final long value, final float base) {
+    public static int log(final long value, final double base) {
         int l = 0;
         for (double i = value; (i >= base); i /= base) {
             ++l;
